test: one home for the root skip, and experiments.tsv its own torn-tail file
Two review findings, both about test files rather than production code.

The root skip was copied into four test files, and two of those copies
landed in this PR: tests/test_appendio.py set 0o200 and asserted the
resulting OSError with no skip at all, so it fails outright in a
container CI that runs as root, while its sibling in the same diff
carried the marker. tests/helpers/rootperms.py is the one home; the
three files in this PR's blast radius import it.

tests/test_append_torn_tail.py was at 796 lines against the 800-line
ratchet, so the next thing added to it had to split it first. The seam
is not arbitrary: every appender left in that file writes JSONL and
every reader of one DROPS a line it cannot parse, while experiments.tsv
is TSV and its reader RENDERS a concatenated line as a run. That is a
different failure with a different fix, so the class moves unchanged
into tests/test_experiments_tsv_torn_tail.py.
